Revert "Add cgroup-bin dependency to our Ubuntu package"
This reverts commit c81bb20f5b2b5d86059c6004e60ba23b03d30fe0.

After re-reading the documentation: "The Recommends field should list packages that would be found together with this one in all but unusual installations."

Thus, "Recommends" is an acceptable place for this dep, and anyone disabling that gets to keep the pieces.

The main crux of why this needs to be reverted is because it breaks Debian completely because "lxc" and "cgroup-bin" can't be installed concurrently.
Upstream-commit: aea7418d8ac73f072a47affcaab04831c2f43bc2
Component: engine