refactor($hack,$docs): be consistent in apt-key keyserver URI usage, as done in #4740
In #4740, the apt-key call in docs is changed to use the keyserver port 80 instead of
port 11371, as the previous call would fail with a restrictive firewall or proxy.
This commit extends the change to all apt-key calls in the repository.

Docker-DCO-1.1-Signed-off-by: Viktor Vojnovski <vojnovski@gmail.com> (github: vojnovski)
Upstream-commit: 4fd82db4beba03a126dfc557c86d5d52e9066dae
Component: engine